PingPong
This game was made by MentosIsGay please don't steal :)
This is a simple pong game I made for fun. Feel free to edit it if you like but please credit me if you're going to publish it.


This was also made with turtle so make sure to have that installed. This works only on python3. If you don't know how to install turtle open windows cmd prompt and type in "pip install turtle"


Thank you :)